feat: creado archivo types para separar la declaracion de enums closes #41
La declaracion de enumerados del archivo usuario se han movido al archivo types, añadiendo al archivo de usuario los import correspondientes
