+++ src/types.ts
@@ -1,0 +1,20 @@
+export enum Rol {
+    Fragger,
+    IGL,
+    Support,
+    Lurker,
+    Awper,
+}
+
+export enum Actitud {
+    Agresivo,
+    Controlador,
+}
+
+export enum Rango {
+    Plata,
+    Oro,
+    Ak,
+    Maestro,
+    EliteGlobal,
+}
