+++ src/usuario.ts
@@ -1,23 +1,4 @@
-export enum Rol {
-    Fragger,
-    IGL,
-    Support,
-    Lurker,
-    Awper,
-}
-
-export enum Actitud {
-    Agresivo,
-    Controlador,
-}
-
-export enum Rango {
-    Plata,
-    Oro,
-    Ak,
-    Maestro,
-    EliteGlobal,
-}
+import {Rol, Actitud, Rango} from './types'
 
 export class Usuario {
     public rol: Rol;
